backend/pyproject.toml
[tool.poetry]
name = "ExcelParser"
version = "0.1.0"
description = "Test project"
authors = ["Vadim Yamanov"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2.6"
django-rest-framework = "^0.1.0"
openpyxl = "^3.0.7"
python-dotenv = "^0.19.0"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
